Start migration to generic data model
